Rename types.d.ts to types.ts
Typescript does not copy declaration files into build dir
