+++ test/ast.d/block-start.ts
@@ -1,6 +1,6 @@
 import * as ast from 'src/ast.js';
 import * as stakr from 'src/stakr.js';
-import { AssembleArg } from 'src/types.d';
+import { AssembleArg } from 'src/types.js';
 import * as _ from 'tap';
 
 await _.test('BlockStart', async (_) => {
