+++ test/ast.d/import-statement.ts
@@ -1,6 +1,6 @@
 import * as ast from 'src/ast.js';
 import * as stakr from 'src/stakr.js';
-import { AssembleArg, LinkArg } from 'src/types.d';
+import { AssembleArg, LinkArg } from 'src/types.js';
 import * as _ from 'tap';
 
 await _.test('ImportStatement', async (_) => {
