+++ test/commands.d/return.ts
@@ -1,6 +1,6 @@
 import { return_ } from 'src/commands.js';
 import * as stakr from 'src/stakr.js';
-import { ExecuteArg } from 'src/types.d';
+import { ExecuteArg } from 'src/types.js';
 import * as _ from 'tap';
 
 const context = new stakr.ExecutionContext();
